Rename BuiltinAlgorithm to ScalaPackage and move to interop as it useful in general, not just for algorithms (e.g., if you have a custom scala package that you want to access)
